#809 changed subutai map to use subutai proxy internally
