+++ cli/map.go
@@ -1,23 +1,14 @@
 package cli
 
 import (
-	"io/ioutil"
-	"net"
 	"os"
-	"strconv"
 	"strings"
-	"time"
-
 	"github.com/subutai-io/agent/config"
 	"github.com/subutai-io/agent/db"
-	"github.com/subutai-io/agent/lib/fs"
-	"github.com/subutai-io/agent/lib/gpg"
-	ovs "github.com/subutai-io/agent/lib/net"
 	"github.com/subutai-io/agent/log"
-	"github.com/nightlyone/lockfile"
 	"path"
-	"github.com/subutai-io/agent/lib/common"
-	"github.com/subutai-io/agent/agent/util"
+	"fmt"
+	"io/ioutil"
 )
 
 var (
@@ -25,118 +16,68 @@ var (
 )
 
 func GetPortMappings(protocol string) []string {
-	return mapList(protocol)
-}
-
-func RemovePortMapping(protocol, sockInt, sockExt, domain string) {
 	protocol = strings.ToLower(protocol)
 
-	if protocol != "tcp" && protocol != "udp" && protocol != "http" && protocol != "https" {
-		log.Error("Unsupported protocol \"" + protocol + "\"")
-	}
-
-	if (protocol == "http" || protocol == "https") && len(domain) == 0 {
-		log.Error("\"-n domain\" is mandatory for http protocol")
-	}
-
-	if protocol == "tcp" || protocol == "udp" {
-		domain = protocol
-	}
-
-	if !ovs.ValidSocket(sockExt) {
-		sockExt = "0.0.0.0:" + sockExt
-	}
-
-	//remove mapping
-	MapRemove(protocol, sockExt, domain, sockInt)
-
-	//restart nginx
-	restart()
-}
-
-func AddPortMapping(protocol, sockInt, sockExt, domain, policy, cert string, sslBackend bool) {
-	protocol = strings.ToLower(protocol)
-
-	if protocol != "tcp" && protocol != "udp" && protocol != "http" && protocol != "https" {
-		log.Error("Unsupported protocol \"" + protocol + "\"")
-	}
-
-	if (protocol == "http" || protocol == "https") && len(domain) == 0 {
-		log.Error("\"-n domain\" is mandatory for http protocol")
-	}
-
-	if protocol == "tcp" || protocol == "udp" {
-		domain = protocol
-	}
-
-	if !ovs.ValidSocket(sockExt) {
-		sockExt = "0.0.0.0:" + sockExt
-	}
-
-	if protocol == "https" && (len(cert) == 0 || !gpg.ValidatePem(cert)) {
-		log.Error("\"-c certificate\" is missing or invalid pem file")
-	}
-
-	if !ovs.ValidSocket(sockInt) {
-		log.Error("Invalid internal socket \"" + sockInt + "\"")
-	}
-
-	if (strings.HasSuffix(sockExt, ":8443") || strings.HasSuffix(sockExt, ":8444") || strings.HasSuffix(sockExt, ":8086")) &&
-		sockInt != "10.10.10.1:"+strings.Split(sockExt, ":")[1] {
-		log.Error("Reserved system ports")
-	}
-
-	//add mapping
-
-	var mapping = protocol + domain + sockInt + sockExt
-	var lock lockfile.Lockfile
-	var err error
-	for lock, err = common.LockFile(mapping, "map"); err != nil; lock, err = common.LockFile(mapping, "map") {
-		time.Sleep(time.Second * 1)
-	}
-	defer lock.Unlock()
-
-	// check sockExt port and create nginx config
-	if portIsNew(protocol, sockInt, domain, &sockExt) {
-		newConfig(protocol, sockExt, domain, cert, sslBackend)
-	}
-
-	// add containers to backend
-	addLine(path.Join(nginxInc, protocol, sockExt+"-"+domain+".conf"),
-		"#Add new host here", "	server "+sockInt+";", false)
-
-	// save information to database
-	saveMapToDB(protocol, sockExt, domain, sockInt)
-	containerMapToDB(protocol, sockExt, domain, sockInt)
-	balanceMethod(protocol, sockExt, domain, policy)
-
-	if socket := strings.Split(sockExt, ":"); socket[0] == "0.0.0.0" {
-		log.Info(ovs.GetIp() + ":" + socket[1])
-	} else {
-		log.Info(sockExt)
-	}
-
-	//restart nginx
-	restart()
-}
-
-func mapList(protocol string) (list []string) {
-	switch protocol {
-	case "tcp", "udp", "http", "https":
-		l, err := db.INSTANCE.PortMapList(protocol)
-		log.Check(log.ErrorLevel, "Reading port mappings from db", err)
-		list = l
-	default:
-		for _, v := range []string{"tcp", "udp", "http", "https"} {
-			l, err := db.INSTANCE.PortMapList(v)
-			if !log.Check(log.ErrorLevel, "Reading port mappings from db", err) {
-				list = append(list, l...)
+	var output []string
+	proxies := GetProxies(protocol)
+	for _, p := range proxies {
+		if protocol == p.Proxy.Protocol || protocol == "" {
+			for _, server := range p.Servers {
+				line := fmt.Sprintf("%s\t%d\t%s\t%s", p.Proxy.Protocol, p.Proxy.Port, server.Socket, p.Proxy.Domain)
+				output = append(output, line)
 			}
 		}
 	}
-	return
+
+	return output
 }
 
+func RemovePortMapping(protocol, domain string, port int, server string) {
+	protocol = strings.ToLower(protocol)
+	domain = strings.ToLower(domain)
+
+	tag := fmt.Sprintf("%s-%d-%s", protocol, port, domain)
+	if protocol == TCP || protocol == UDP {
+		tag = fmt.Sprintf("%s-%d-stream", protocol, port)
+	}
+
+	if server != "" {
+		RemoveProxiedServer(tag, server)
+	} else {
+		RemoveProxy(tag)
+		return
+	}
+
+	servers, err := db.FindProxiedServers(tag, "")
+	log.Check(log.ErrorLevel, "Getting proxied servers from db", err)
+
+	if len(servers) == 0 {
+		RemoveProxy(tag)
+	}
+}
+
+func AddPortMapping(protocol, domain, loadBalancing string, port int, server, certPath string, redirect80Port, sslBackend bool) {
+	protocol = strings.ToLower(protocol)
+	domain = strings.ToLower(domain)
+
+	tag := fmt.Sprintf("%s-%d-%s", protocol, port, domain)
+	if protocol == TCP || protocol == UDP {
+		tag = fmt.Sprintf("%s-%d-stream", protocol, port)
+	}
+
+	proxy, err := db.FindProxyByTag(tag)
+	log.Check(log.ErrorLevel, "Getting proxy from db", err)
+
+	if proxy == nil {
+		CreateProxy(protocol, domain, loadBalancing, tag, port, redirect80Port, sslBackend, certPath)
+		proxy, err = db.FindProxyByTag(tag)
+		log.Check(log.ErrorLevel, "Getting proxy from db", err)
+	}
+
+	AddProxiedServer(tag, server)
+}
+
+//todo remove after one version
 func MapRemove(protocol, sockExt, domain, sockInt string) {
 	log.Debug("Removing mapping: " + protocol + " " + sockExt + " " + domain + " " + sockInt)
 
@@ -156,8 +97,8 @@ func MapRemove(protocol, sockExt, domain, sockInt string) {
 				}
 				os.Remove(path.Join(nginxInc, protocol, sockExt+"-"+domain+".conf"))
 				if protocol == "https" {
-					os.Remove(path.Join(webSslPath, "https-"+sockExt+"-"+domain+".key"))
-					os.Remove(path.Join(webSslPath, "https-"+sockExt+"-"+domain+".crt"))
+					os.Remove(path.Join(selfSignedCertsDir, "https-"+sockExt+"-"+domain+".key"))
+					os.Remove(path.Join(selfSignedCertsDir, "https-"+sockExt+"-"+domain+".crt"))
 				}
 			}
 		}
@@ -167,176 +108,9 @@ func MapRemove(protocol, sockExt, domain, sockInt string) {
 		}
 		os.Remove(path.Join(nginxInc, protocol, sockExt+"-"+domain+".conf"))
 		if protocol == "https" {
-			os.Remove(path.Join(webSslPath, "https-"+sockExt+"-"+domain+".key"))
-			os.Remove(path.Join(webSslPath, "https-"+sockExt+"-"+domain+".crt"))
+			os.Remove(path.Join(selfSignedCertsDir, "https-"+sockExt+"-"+domain+".key"))
+			os.Remove(path.Join(selfSignedCertsDir, "https-"+sockExt+"-"+domain+".crt"))
 		}
-	}
-}
-
-func isFree(protocol, sockExt string) bool {
-	switch protocol {
-	case "tcp", "http", "https":
-		if ln, err := net.Listen("tcp", sockExt); err == nil {
-			ln.Close()
-			return true
-		}
-	case "udp":
-		if addr, err := net.ResolveUDPAddr("udp", sockExt); err == nil {
-			if ln, err := net.ListenUDP("udp", addr); err == nil {
-				ln.Close()
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func portIsNew(protocol, sockInt, domain string, sockExt *string) bool {
-	socket := strings.Split(*sockExt, ":")
-	if len(socket) > 1 && socket[1] != "" {
-		if port, err := strconv.Atoi(socket[1]); err != nil || port < 1000 || port > 65536 {
-			if !(strings.Contains(protocol, "http") && (port == 80 || port == 443)) {
-				log.Error("Port number in \"external\" should be integer in range of 1000-65536")
-			}
-		}
-		if isFree(protocol, *sockExt) {
-			return true
-		}
-
-		//for non-80 port dont allow multiple mappings
-		if !checkPort(protocol, *sockExt, "", "") && socket[1] != "80" {
-			log.Error("Port is busy")
-		} else if checkPort(protocol, *sockExt, domain, sockInt) {
-			log.Error("Mapping already exists")
-		}
-		//just check to see if config files for the same external port and diff internal ports already exists
-		return !checkPort(protocol, *sockExt, domain, "")
-	}
-	for port := strconv.Itoa(common.RandomInt(1000, 65536)); isFree(protocol, socket[0]+":"+port); port = strconv.Itoa(common.RandomInt(1000, 65536)) {
-		*sockExt = socket[0] + ":" + port
-		return true
-	}
-	return false
-}
-
-func newConfig(protocol, sockExt, domain, cert string, sslbcknd bool) {
-	log.Check(log.WarnLevel, "Creating nginx include folder",
-		os.MkdirAll(path.Join(nginxInc, protocol), 0755))
-	conf := path.Join(nginxInc, protocol, sockExt+"-"+domain+".conf")
-
-	switch protocol {
-	case "https":
-		log.Check(log.ErrorLevel, "Creating certificate dirs", os.MkdirAll(webSslPath, 0755))
-		fs.Copy(path.Join(conftmpl, "vhost-ssl.example"), conf)
-		addLine(conf, "return 301 https://$host$request_uri;  # enforce https", "	    return 301 https://$host:"+strings.Split(sockExt, ":")[1]+"$request_uri;  # enforce https", true)
-		addLine(conf, "listen	443;", "	listen "+sockExt+";", true)
-		addLine(conf, "server_name DOMAIN;", "	server_name "+domain+";", true)
-		if sslbcknd {
-			addLine(conf, "proxy_pass http://DOMAIN-upstream/;", "	proxy_pass https://https-"+strings.Replace(sockExt, ":", "-", -1)+"-"+domain+";", true)
-		} else {
-			addLine(conf, "proxy_pass http://DOMAIN-upstream/;", "	proxy_pass http://https-"+strings.Replace(sockExt, ":", "-", -1)+"-"+domain+";", true)
-		}
-		addLine(conf, "upstream DOMAIN-upstream {", "upstream https-"+strings.Replace(sockExt, ":", "-", -1)+"-"+domain+" {", true)
-
-		crt, key := util.ParsePem(cert)
-		log.Check(log.WarnLevel, "Writing certificate body", ioutil.WriteFile(path.Join(webSslPath, "https-"+sockExt+"-"+domain+".crt"), crt, 0644))
-		log.Check(log.WarnLevel, "Writing key body", ioutil.WriteFile(path.Join(webSslPath, "https-"+sockExt+"-"+domain+".key"), key, 0644))
-
-		addLine(conf, "ssl_certificate "+path.Join(webSslPath, "UNIXDATE.crt;"),
-			"ssl_certificate "+path.Join(webSslPath, "https-"+sockExt+"-"+domain+".crt;"), true)
-		addLine(conf, "ssl_certificate_key "+path.Join(webSslPath, "UNIXDATE.key;"),
-			"ssl_certificate_key "+path.Join(webSslPath, "https-"+sockExt+"-"+domain+".key;"), true)
-	case "http":
-		fs.Copy(path.Join(conftmpl, "vhost.example"), conf)
-		addLine(conf, "listen 	80;", "	listen "+sockExt+";", true)
-		addLine(conf, "return 301 http://$host$request_uri;", "	    return 301 http://$host:"+strings.Split(sockExt, ":")[1]+"$request_uri;", true)
-		addLine(conf, "server_name DOMAIN;", "	server_name "+domain+";", true)
-		addLine(conf, "proxy_pass http://DOMAIN-upstream/;", "	proxy_pass http://http-"+strings.Replace(sockExt, ":", "-", -1)+"-"+domain+";", true)
-		addLine(conf, "upstream DOMAIN-upstream {", "upstream http-"+strings.Replace(sockExt, ":", "-", -1)+"-"+domain+" {", true)
-		if !strings.HasSuffix(sockExt, ":80") {
-			httpRedirect(sockExt, domain)
-		}
-	case "tcp":
-		fs.Copy(path.Join(conftmpl, "stream.example"), conf)
-		addLine(conf, "listen PORT;", "	listen "+sockExt+";", true)
-	case "udp":
-		fs.Copy(path.Join(conftmpl, "stream.example"), conf)
-		addLine(conf, "listen PORT;", "	listen "+sockExt+" udp;", true)
-	}
-	addLine(conf, "server localhost:81;", " ", true)
-	addLine(conf, "upstream PROTO-PORT {", "upstream "+protocol+"-"+strings.Replace(sockExt, ":", "-", -1)+"-"+domain+" {", true)
-	addLine(conf, "proxy_pass PROTO-PORT;", "	proxy_pass "+protocol+"-"+strings.Replace(sockExt, ":", "-", -1)+"-"+domain+";", true)
-}
-
-func balanceMethod(protocol, sockExt, domain, policy string) {
-	replaceString := "upstream " + protocol + "-" + strings.Replace(sockExt, ":", "-", -1) + "-" + domain + " {"
-	replace := false
-
-	if !checkPort(protocol, sockExt, domain, "") {
-		log.Error("Port is not mapped")
-	}
-	switch policy {
-	case "round-robin", "round_robin":
-		policy = "#round-robin"
-		//  "least_conn":
-	case "least_time":
-		if protocol == "tcp" {
-			policy = policy + " connect"
-		} else {
-			policy = policy + " header"
-			log.Warn("This policy is not supported in http upstream")
-			return
-		}
-	case "hash":
-		policy = policy + " $remote_addr"
-	case "ip_hash":
-		if protocol != "http" {
-			log.Warn("ip_hash policy allowed only for http protocol")
-			return
-		}
-	default:
-		log.Debug("Unsupported balancing method \"" + policy + "\", ignoring")
-		return
-	}
-
-	p, err := db.INSTANCE.GetBalancingPolicy(protocol, sockExt, domain)
-	log.Check(log.ErrorLevel, "Reading port mapping from db", err)
-
-	if len(p) != 0 && p != policy {
-		replaceString = "; #policy"
-		replace = true
-	} else if p == policy {
-		return
-	}
-	log.Check(log.ErrorLevel, "Saving map method", db.INSTANCE.SetBalancingPolicy(protocol, sockExt, domain, policy))
-
-	addLine(path.Join(nginxInc, protocol, sockExt+"-"+domain+".conf"),
-		replaceString, "	"+policy+"; #policy", replace)
-}
-
-func httpRedirect(sockExt, domain string) {
-	var redirect = `server {
-	    listen      80; #redirect
-    	server_name ` + domain + `;
-    	return 301 http://$host:` + strings.Split(sockExt, ":")[1] + `$request_uri;
-}`
-
-	addLine(path.Join(nginxInc, "http", sockExt+"-"+domain+".conf"),
-		"#redirect placeholder", redirect, true)
-
-}
-
-func saveMapToDB(protocol, sockExt, domain, sockInt string) {
-	if !checkPort(protocol, sockExt, domain, sockInt) {
-		log.Check(log.ErrorLevel, "Saving port map to database", db.INSTANCE.PortMapSet(protocol, sockExt, domain, sockInt))
-	}
-}
-
-func containerMapToDB(protocol, sockExt, domain, sockInt string) {
-	list, err := db.INSTANCE.GetContainerByKey("ip", strings.Split(sockInt, ":")[0])
-	log.Check(log.ErrorLevel, "Reading container metadata from db", err)
-	for _, name := range list {
-		log.Check(log.ErrorLevel, "Saving port mapping to db", db.INSTANCE.SaveContainerPortMapping(name, protocol, sockExt, domain, sockInt))
 	}
 }
 
@@ -350,4 +124,31 @@ func deletePortMap(protocol, sockExt, domain, sockInt string) int {
 	left, err := db.INSTANCE.PortMapDelete(protocol, sockExt, domain, sockInt)
 	log.Check(log.ErrorLevel, "Removing port mapping from db", err)
 	return left
+}
+
+// addLine adds, removes, replaces and checks if line exists in specified file
+func addLine(path, after, line string, replace bool) bool {
+	f, err := ioutil.ReadFile(path)
+	if !log.Check(log.DebugLevel, "Cannot read file "+path, err) {
+		lines := strings.Split(string(f), "\n")
+		for k, v := range lines {
+			if strings.Contains(v, after) {
+				if line != "" {
+					if replace {
+						log.Debug("Replacing " + lines[k] + " with " + line)
+						lines[k] = line
+					} else {
+						log.Debug("Adding " + line + " after " + lines[k])
+						lines[k] = after + "\n" + line
+					}
+				} else {
+					return true
+				}
+			}
+		}
+		str := strings.Join(lines, "\n")
+		log.Check(log.FatalLevel, "Writing new proxy config",
+			ioutil.WriteFile(path, []byte(str), 0744))
+	}
+	return false
 }
