+++ lib/common/util.go
@@ -10,8 +10,6 @@ import (
 	"strings"
 	"os"
 	"path"
-	"time"
-	"math/rand"
 )
 
 func RunNRecover(g func()) {
@@ -54,14 +52,4 @@ func LockFile(name string, command string) (lockfile.Lockfile, error) {
 	}
 
 	return lock, nil
-}
-
-func RandomInt(min, max int) int {
-	rand.Seed(time.Now().Unix())
-	return rand.Intn(max-min) + min
-}
-
-
-func IsZeroOfUnderlyingType(x interface{}) bool {
-	return x ==nil || reflect.DeepEqual(x, reflect.Zero(reflect.TypeOf(x)).Interface())
 }
